refactor track save, move into own source
