+++ openrct2.vcxproj
@@ -182,6 +182,7 @@
     <ClCompile Include="src\ride\track.c" />
     <ClCompile Include="src\ride\track_data.c" />
     <ClCompile Include="src\ride\track_design.c" />
+    <ClCompile Include="src\ride\track_design_save.c" />
     <ClCompile Include="src\ride\track_list.c" />
     <ClCompile Include="src\ride\track_paint.c" />
     <ClCompile Include="src\ride\transport\chairlift.c" />
